nfs: add advertise endpoints
We use Kubernetes Services to expose the NFS server to the clients. To this
end, we have a central "Service" resource that represents a Service IP that
has one or more ports. Every NFS export gets its own Port. To connect the two,
Kubernetes uses "EndpointSlice" resource, which we create and remove whenever
the DRBD resource gets promoter or demoted.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>