snapshots: use URL-based snapshot name
In order to prepare for group snapshots, we need a better "Snapshot ID" to
uniquely identify a snapshot of a specific volume. Since the ID can be
basically anything, we can make it a rich url:

  <snapshot-type>://<remote>/<source-resource>/<snapshot-name>

With this scheme, a group snapshot still results in unique snapshot IDs, as
the ID contains the source volume name. It also helps us differentiate between
"local-only" snapshots and backups to LINSTOR remotes by having the type and
remote name directly in the ID.

One wrinkle is that we:
* Have to support existing snapshots that have not been provisioned with this
  scheme in mind.
* Have to be able to recover from incomplete snapshot creations. That means
  that we retry the Creation of the snapshot with the same initial snapshot
  name.

Luckily, the solution is the same for both cases: in case we get a non-url
snapshot ID, we fall back to the existing search for snapshots, assuming that
it is unique.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>