update golinstor client
This allows us to remove the small gap when cloning a volume that could
lead to an incomplete volume being provisioned. New LINSTOR versions
can remove properties on the initial clone request.

Also remove dependency on golang.org/x/exp, as all the necessary calls
are now part of stdlib.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>