ci: sign images
Enable keyless signing of images using cosign. The signature is tied to the
github action, so as long as people trust github runners, they know that
the image is actually the one built by this workflow.

Since we use docker buildx bake, to build two different images, we need to
extract the exact repository names and digest from the build metadata. This
leads to a pretty ugly jq expression, but is known to work[1].

[1]: https://github.com/piraeusdatastore/piraeus/blob/master/.github/workflows/build-docker.yaml#L57-L61

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>