nfs: fix owner reference for endpoints
Setting BlockOwnerDeletion=true on the OwnerReference for the endpoint slice
was intended to block removal of the Pod running the NFS instance until all
service references have been removed.

This however would require additional privileges on the NFS server Pod, as
Kubernetes verifies that the account creating the EndpointSlice is allowed
to set finalizers on the Pod.

Since we don't really need block deletion of the Pod until the EndpointSlice
is removed, we simply remove the BlockOwnerDeletion.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>