snapshots: search other remotes if not found
With the introduction of URL-based snapshot names, we removed the possibility
to have snapshots work between clusters when the remote name did not match.
A snapshot with the name:

S3://my-remote/res1/snap1

Could not be used from a cluster where "my-remote" was instead named
"remote-from-another-cluster".

This came as a surprise to some users, so we make the effort to search
additional remotes if the named remote does not contain the snapshot, or even
does not exist at all.

Fixes: 8edb9d331010 ("snapshots: use URL-based snapshot name")
Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>