Fix behaviour of "delete-local" snapshots
When creating remove snapshots with the "delete-local" parameter set, we should
delete local snapshot, but not the remote backup.

However, with the change to use URL-based snapshot names, we accidentally
changed the behaviour by passing in the full snapshot name, including the
remote part. This caused the routine to delete also the backup on the
remote location.

The fix is to implement what was still mentioned in the comment above: create
a temporary copy of the SnapshotId without the remote part set.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>
Fixes: 8edb9d331010 ("snapshots: use URL-based snapshot name")