fix compatibility with LINSTOR 1.33 backup API
LINSTOR 1.33.0 changed the backup API, so that a successful up- or download
is no longer indicated by a flag on the snapshot resource. Instead some
properties are used in order to support shipping to multiple targets.

This meant our existing implementation that checks if a LINSTOR snapshot is
actually a "remote" snapshot broke. To fix this, we first check for the new
properties on the snapshot and use those as basis for the snapshot remote
names and readiness.

LINSTOR now allows shipping to more than one target, but we only expect one
"remote" location to exist, we need to select the correct one. In our case,
this should always be the earliest remote that started shipping, so if there
are multiple possible destination, we report the one that first started
uploading.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>