Fix: include node name in topology segments for RemoteAccessPolicyLocalOnly
When using allowRemoteVolumeAccess: false, the CSI driver should set
nodeAffinity on the PV to restrict pods to nodes with local replicas.

The RemoteAccessPolicyLocalOnly policy uses topology.LinstorNodeKey
(linbit.com/hostname) to filter accessible nodes. However, the node name
was not included in the topology segments passed to AccessibleSegments(),
causing it to return empty segments and thus no nodeAffinity restriction.

This fix adds the node name to the segments map so that
RemoteAccessPolicyLocalOnly can properly identify which nodes have
diskful replicas and set the appropriate nodeAffinity on the PV.

Signed-off-by: Christian Wohlert <wohlert@appbase.hamburg>